Setting up environment, restructuring files
