+++ main.py
@@ -1,0 +1,3 @@
+import os
+
+os.system("python src/train_model.py")
