+++ tests/general_test.py
@@ -1,0 +1,7 @@
+
+def inc(x):
+    return x + 2
+
+
+def test_answer():
+    assert inc(3) == 5
